change: merge EventsPage and TimersPage into HomePage
