+++ src/components/Meta/Meta.jsx
@@ -5,6 +5,8 @@ import { useEffect } from "react";
 //import styles from "./Meta.module.css";
 import { setPageMeta } from "@stores/metaStore";
 
+const prefix = " - FGO Timers";
+
 export default function Meta({
   title,
   description,
@@ -24,14 +26,14 @@ export default function Meta({
 
   return (
     <Head>
-      <title>{`${title}${noTitleSuffix ? "" : " - FGO Tools"}`}</title>
+      <title>{`${title}${noTitleSuffix ? "" : prefix}`}</title>
       <meta
         name="twitter:title"
-        content={`${title}${noTitleSuffix ? "" : " - FGO Tools"}`}
+        content={`${title}${noTitleSuffix ? "" : prefix}`}
       />
       <meta
         property="og:title"
-        content={`${title}${noTitleSuffix ? "" : " - FGO Tools"}`}
+        content={`${title}${noTitleSuffix ? "" : prefix}`}
       />
       <meta property="og:site_name" content="FGO Tools" />
       <meta name="description" content={description} />
