+++ src/styles/HomePage.module.css
@@ -1,29 +1,10 @@
-.button {
-  svg {
-    fill: theme(colors.primary);
-  }
-  &:hover {
-    svg {
-      fill: theme(colors.hover);
-    }
-  }
-  &.discord:hover {
-    &:hover svg {
-      fill: #5865f2 !important;
-    }
-  }
+.grid {
+  display: flex;
+  justify-content: space-around;
+  flex-wrap: wrap;
+  gap: [0.66em, null, 1.33em];
 }
 
-.section > img {
-  width: 256px;
-  max-width: [75%, 90%];
-  height: auto;
-  display: block;
-  margin: 1.66em auto;
-}
-.section > h1 {
-  font-size: [1.5em, 2em, 3em];
-}
 .section > p {
   font-size: [0.8em, null, 0.66em];
 }
