+++ src/utils/navRoutes.js
@@ -1,6 +1,4 @@
 export const navRoutes = [
-  { link: "/", text: "Home", test: /^\/$/ },
-  { link: "/timers", text: "Timers" },
-  { link: "/events", text: "Events", test: /^\/events\/?$/ },
+  { link: "/", text: "Timers", test: /^\/$/ },
   { link: "/upgrades", text: "Upgrades" }
 ];
